chore(tile-debug): refactoring
