+++ src/utils/bounding-volume.ts
@@ -1,4 +1,5 @@
 import type { ObbData } from './types';
+
 import { Tile3D } from "@loaders.gl/tiles";
 import { OrientedBoundingBox, BoundingSphere } from "@math.gl/culling";
 import { OBB, MBS } from "../constants/bounding-volumes";
@@ -8,7 +9,9 @@ import { OBB, MBS } from "../constants/bounding-volumes";
  * @param tile
  * @param boundingType
  */
-export const createBoundingVolumeFromTile = (tile: Tile3D, boundingType: string): BoundingSphere | OrientedBoundingBox => {
+export const createBoundingVolumeFromTile = (tile: Tile3D): BoundingSphere | OrientedBoundingBox => {
+  const boundingType = getBoundingType(tile);
+
   switch (boundingType) {
     case OBB: {
       return createBoundingBoxFromTileObb(tile.header.obb);
@@ -19,6 +22,17 @@ export const createBoundingVolumeFromTile = (tile: Tile3D, boundingType: string)
     default:
       throw new Error("Validator - Not supported Bounding Volume Type");
   }
+};
+
+/**
+ * Defines the Bounding Box type
+ * @param tile
+ */
+export const getBoundingType = (tile: Tile3D): string => {
+  if (tile.header.obb || tile.boundingVolume instanceof OrientedBoundingBox) {
+    return OBB;
+  }
+  return MBS;
 };
 
 /**
