+++ src/utils/normals-utils.ts
@@ -1,4 +1,5 @@
 import type { Tile3D } from "@loaders.gl/tiles";
+
 import { Vector3, Matrix4 } from "@math.gl/core";
 import { Ellipsoid } from "@math.gl/geospatial";
