+++ src/utils/validation-utils/attributes-validation/geometry-vs-texture-metrics.ts
@@ -1,7 +1,7 @@
 import type { Tile3D } from "@loaders.gl/tiles";
-import type { GeometryVSTextureMetrics } from './types';
+import type { GeometryVSTextureMetrics } from '../../types';
 
-import { getTriangleVertices, getTriangleArea } from './triangles-calculation';
+import { getTriangleVertices, getTriangleArea } from "./triangles-calculation";
 
 /**
  * Generates geometry vs texture metrics
